refactor: moving food item related components inside separate folder
rename components to be more specific with `FoodItem`
