+++ src/components/FoodItems/ModifyFoodItem/ModifyFoodItemPage.tsx
@@ -7,14 +7,14 @@ import {
   IonToolbar,
   useIonToast,
 } from "@ionic/react";
-import { api } from "../../utils/api";
-import ModifyItemForm from "./ModifyItemForm";
+import { api } from "../../../utils/api";
+import ModifyFoodItemForm from "./ModifyFoodItemForm";
 import type { FC } from "react";
 import React from "react";
 import type { FoodItem } from "@prisma/client";
-import { capitalizeFirstLetter } from "../../utils/string";
+import { capitalizeFirstLetter } from "../../../utils/string";
 
-const ModifyItemPage: FC<{
+const ModifyFoodItemPage: FC<{
   container: "fridge" | "pantry";
   foodItem?: FoodItem;
   onClose: () => void;
@@ -75,7 +75,7 @@ const ModifyItemPage: FC<{
           </IonTitle>
         </IonToolbar>
       </IonHeader>
-      <ModifyItemForm
+      <ModifyFoodItemForm
         initialName={foodItem?.name}
         initialExpirationDate={foodItem?.expirationDate.toISOString()}
         onSave={saveHandler}
@@ -85,4 +85,4 @@ const ModifyItemPage: FC<{
   );
 };
 
-export default ModifyItemPage;
+export default ModifyFoodItemPage;
